change io_timeout_ type from uint32_t to unsigned long

+++ sources/ffmpeg_demuxer_source/src/stsw_ffmpeg_demuxer.cc
@@ -63,7 +63,7 @@ FFmpegDemuxer::~FFmpegDemuxer()
 }
 int FFmpegDemuxer::Open(const std::string &input, 
              const std::string &ffmpeg_options_str,
-             uint32_t io_timeout, 
+             unsigned long io_timeout, 
              int play_mode)
 {
     using namespace stream_switch;
